♻️ Refactor: GridFooter 컴포넌트 분리 및 목업파일 불필요한 주석 제거 (#61)
