+++ src/mocks/data/idols.ts
@@ -1,5 +1,3 @@
-// src/mocks/data/idols.ts
-
 // ==============================
 // 타입
 // ==============================
